Release 0.9.5
Moved Crusher Recipes to the config file to allow a mod-pack creator to alter the default recipes. (detailed info is in the config file)
Added Custom Crusher Recipes to the config file to allow a mod-pack creator to add extra recipes recipes. (detailed info is in the config file)
Added a "Speed Upgrade module for the Coal Powered Crusher.
Added "goforlaunch_mode", toggle-able in the config file, to change to a different way the mod drops it's contents. "GFL Mode" is custom made for the "Go For Launch(upcoming)" mod-pack. This mode let's all ores drop a Dirty Chunk which can be smelted to an ingot or a mod-pack creator finds a way to make 2 Clean Chunks with some machine. They smelt into 1 ingot as well but this gives you another way to duplicate "Ores" (as being asked for by the GFL Pack creator). When this mode is enabled it creates the 2 new "Chunks" on startup, if disabled the "Chunks" are NOT registered in-game as they provide no use for the "Normal" (current) ways this mod works.

N.B.: By updating to this new version nothing changes except for the Crusher Recipes being put in the config file to let you alter it or add stuff. The Mod will still continue to function as it does now. The GFL Mode is off by default but can be activated.